# Map Data checkbox: perhaps use toggle slider instead