… vendor/assets/leaflet/leaflet.osm.js