@AntonKhorev 2024-07-20
Are you planning to get the changes to `vendor/assets/leaflet/leaflet.osm.js` merged in https://github.com/openstreetmap/leaflet-osm ?