@nertc 2024-07-24
I created a PR for the [leaflet-osm](https://github.com/openstreetmap/leaflet-osm) (https://github.com/openstreetmap/leaflet-osm/pull/38) that needs to be merged before this PR will be, as asynchronous loading part depends on it.