@nertc 2024-09-03
This PR is linked to the https://github.com/openstreetmap/leaflet-osm/pull/38. If leaflet-osm PR is not merged, this PR can't be merged.