@nertc 2024-09-23
PR was updated to include changes made in the https://github.com/openstreetmap/leaflet-osm/pull/38. As leaflet-osm PR is now merged, this PR can be reviewed and merged.