@AntonKhorev 2024-09-23
Now you can check *Map Data*, wait for the rendering to start then quickly click *Map Data* 2*n times and watch how data is added then removed then added then removed... n times. While this is happening and *Map Data* is checked, pan to a different place to get a stream of javascript errors.