@nertc 2024-10-01
@AntonKhorev I couldn't reproduce JS errors in the console. But I changed logic and now when user clicks "Map Data" checkbox, previous action is skipped immediately. Therefore, there will be no 2*n loads and errors. When we will be okay with the solution, I'll create a new PR in the [leaflet-osm](https://github.com/openstreetmap/leaflet-osm).