@nertc 2024-10-23
If there are no other change requests for this PR, I'll create a new PR for the [leaflet-osm](https://github.com/openstreetmap/leaflet-osm) to merge changes done in the `vendor/assets/leaflet/leaflet.osm.js`.