@nertc 2025-02-04
PR was updated. Branch was synced with master and conflicts were resolved. Also, instead of handling canceling layer loading in the `setTimeout` callback, now it is handled by clearing `setTimeouts`.

Optimizations for layer rendering were done in https://github.com/openstreetmap/leaflet-osm/pull/43, but still this PR is relevant, because it adds loading spinner for those cases when data is processed by the back end, and also for slower CPUs (for example, when I set `6x slowdown` to the CPU) it will provide smooth asynchronous loading without any freezes.