@nertc 2025-03-10
PR was updated. `master` was rebased. Conflicts were resolved. Failed tests were not impacted by this PR (for more information see https://github.com/openstreetmap/openstreetmap-website/pull/5439#issuecomment-2711306484).