@nertc 2025-03-10
This PR is linked to the https://github.com/openstreetmap/leaflet-osm/pull/47.