@nertc 2025-03-12
PR was updated. `master` was rebased.

@tomhughes Thank you for the review. Spinner issue on page refresh was fixed.