@deevroman 2025-03-14
This PR brought two problems:

1. When moving the map, a white flash occurs and all objects on the map are redrawn:

https://github.com/user-attachments/assets/3c87c769-174c-4513-a5be-115eb9517bb8

2. Rendering slowed down. What used to occupy ~400ms now occupies ~1.3s. Hiding a data layer was previously performed on ~100ms, now ~1.5s (!)

Both problems arise due to the fact that before that the redrawing of the map was performed only once, after the completion of the removal/adding of objects to the SVG-element of the map.

I can agree that the first appearance of the data layer looks beautiful. But the work with the map worsened.