@nertc 2025-03-19
@deevroman Thank you for your feedback. #5799 disabled asynchronous layer loading and, therefore, there shouldn't be any flashes or slow rendering process anymore.