@deevroman 2025-03-19
@nertc no, https://github.com/openstreetmap/openstreetmap-website/pull/5799 disabled asynchrony not for the Map Data layer, but for the layer that shows the active object.