@deevroman 2025-03-19
What do you use to test CPU slowdowns? Test rendering >4k features.

Can you give us an example of when the user has time to notice the browser hanging, while making the user guess that they are waiting for rendering, and not downloading data?