@nertc 2025-03-20
For `CPU slowdown` I use browser built-in functionality.

Difference between getting data from server and rendering them is that while getting data, there are no CPU intense tasks and loading icon is shown next to the `data layer` checkbox. When rendering is started, it uses CPU to render data and may freeze application or browser.

In the video you can see when rendering data can take long time (full video can't be uploaded because of the limitations of GitHub, in reality it took about 30 seconds to render all elements). In such case, there would be 30 seconds freeze of browser if there was synchronous process. 

https://github.com/user-attachments/assets/266a42bc-f150-42e0-ae69-709781c7bda0