@deevroman 2025-03-20
> For CPU slowdown I use browser built-in functionality.

Ah, chromium, got it. 20x slowdown is unrealistic. With this value, I can't move the map normally, even without the data layer. But even with 20x slowdown, the layer with 11k features is drawn in 4 seconds when rendered synchronously on my PC.

> In such case, there would be 30 seconds freeze of browser if there was synchronous process.

This is not true because synchronous rendering is faster than asynchronous rendering. 

In general, I will most likely suggest a PR that disables asynchrony.